+++ pkg/pt/lbvh.go
@@ -3,20 +3,22 @@ package pt
 import (
 	"math"
 	"runtime"
-	"sort"
 	"sync"
 )
 
 var MORTON_SIZE = uint32(math.Pow(2, 20))
+
+const BUCKET_COUNT = 4096
 
 func DefaultLBVH(prims []tracable) BVH {
 	return LBVH(prims, enclosing(prims), runtime.GOMAXPROCS(0))
 }
 
 func LBVH(prims []tracable, enclosing aabb, threads int) BVH {
-	paris := assignMortonCodes(prims, enclosing, MORTON_SIZE, threads)
-	sortPairs(paris, threads)
-	bvh := constructLBVH(paris, MORTON_SIZE, threads)
+	pairs := assignMortonCodes(prims, enclosing, MORTON_SIZE, threads)
+	maxMorton := uint64(math.Pow(float64(MORTON_SIZE), 3))
+	pairs = sortMortonPairs(pairs, BUCKET_COUNT, maxMorton, threads)
+	bvh := constructLBVH(pairs, MORTON_SIZE, threads)
 	bvh.prims = prims
 	bvh.storeLeaves()
 	bvh.updateBounding(threads)
@@ -28,6 +30,8 @@ type mortonPair struct {
 	mortonCode uint64
 }
 
+// Iterates over all primitives in parallel and assigns morton codes
+// TODO: Check if returning a corresponding morton code slice is faster than using structs!
 func assignMortonCodes(prims []tracable, enclosing aabb, mortonSize uint32, threads int) []mortonPair {
 	pairs := make([]mortonPair, len(prims))
 	batchSize := int(math.Ceil(float64(len(prims)) / float64(threads)))
@@ -59,6 +63,7 @@ func assignMortonCodes(prims []tracable, enclosing aabb, mortonSize uint32, thre
 	return pairs
 }
 
+// Computes a morton code according to the quantized primitive centroid
 func computeMorton(prim tracable, morton *Morton, enclosing aabb, mortonSize uint32) uint64 {
 	center := prim.bounding().barycenter
 	deltaX := math.Abs(enclosing.bounds[0].X - center.X)
@@ -72,14 +77,7 @@ func computeMorton(prim tracable, morton *Morton, enclosing aabb, mortonSize uin
 	return encoded
 }
 
-// Parallel Bucket sort according to https://www.sjsu.edu/people/robert.chun/courses/cs159/s3/N.pdf
-func sortPairs(pairs []mortonPair, threads int) {
-	// TODO: Implement some parallel sorting algorithm
-	sort.Slice(pairs, func(i, j int) bool {
-		return pairs[i].mortonCode < pairs[j].mortonCode
-	})
-}
-
+// Constructs BVH by inserting sorted morton primitive pairs into a binary radix tree
 func constructLBVH(pairs []mortonPair, mortonSize uint32, threads int) BVH {
 	splitMask := uint64(math.Pow(float64(mortonSize), 3)) / 2
 	wg := sync.WaitGroup{}
@@ -89,6 +87,7 @@ func constructLBVH(pairs []mortonPair, mortonSize uint32, threads int) BVH {
 		wg:   &wg,
 	}
 
+	// Start workers, each worker will find a split in its given interval and spawn 2 new jobs
 	for i := 0; i < threads; i++ {
 		go func(q *lbvhWorkerQueue) {
 			for job := range queue.jobs {
@@ -148,9 +147,11 @@ func (job *lbvhJob) process(queue *lbvhWorkerQueue) {
 		return
 	}
 
+	// Find the split in the given interval where the most significant bit first changes
 	splitIndex := findSplit(job.pairs, job.splitMask)
 
 	// TODO: Remove and rollup singletons after?
+	// If there is no split, only spawn one job, which makes pruning step afterwards obsolete
 	if splitIndex == 0 || splitIndex == len(job.pairs) {
 		queue.add(&lbvhJob{
 			pairs:      job.pairs,
@@ -160,7 +161,7 @@ func (job *lbvhJob) process(queue *lbvhWorkerQueue) {
 		})
 		return
 	}
-	// New Branch
+	// Create a new branch and spawn new jobs for both children
 	branch := newBranch(2)
 	job.parent.addChild(branch, job.childIndex)
 	left := job.pairs[:splitIndex]
@@ -192,7 +193,6 @@ func findSplit(pairs []mortonPair, splitMask uint64) int {
 	r := len(pairs) - 1
 	for l <= r {
 		m := (l + r) / 2
-		// TODO: Size comparison more efficient than bit mask?
 		if (pairs[m].mortonCode & splitMask) == 0 {
 			// Continue search in right half
 			l = m + 1