+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented PHR BVH builder
 - benchmark package including all benchmarks conducted
 - Added optional unlit renderer
+- Implemented parallel bucket sort
 ### Changed 
 - Fixed bug in enclosing functions
 - Parallelized PHR find initial Cut